restructure and add windows functionality
